Create C_Asteriod_Location_Project
This program asks the user for their credentials, and after verifying that they 
are correct with fileIO it proceeds to create a random array of asteroids, calculate their distance from the user(0,0), and print the X, Y, & Z points in an asteroid file and these along with the distance in a distance file. After doing so the program warns the user of asteroid proximity, and graphs the distance of each asteroid.